# apps/api/openapi.generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TB Group API",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
...
  "paths": {
...
    "/api/banners": {
      "get": {
        "summary": "List /api/banners",
        "responses": {
          "200": {
            "description": "List",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "placement": {
                            "type": "string",
                            "enum": [
                              "HOME_HERO",
                              "CTA_PRIMARY",
                              "CTA_SECONDARY"
                            ]
                          },
                          "title": {
                            "type": "string"
                          },
                          "subtitle": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "ctaLabel": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "ctaLink": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "serviceId": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "order": {
                            "type": "number"
                          },
                          "isActive": {
                            "type": "boolean"
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "placement",
                          "title",
                          "subtitle",
                          "ctaLabel",
                          "ctaLink",
                          "serviceId",
                          "order",
                          "isActive",
                          "createdAt",
                          "updatedAt"
                        ]
                      }
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create /api/banners",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "placement": {
                    "type": "string",
                    "enum": [
                      "HOME_HERO",
                      "CTA_PRIMARY",
                      "CTA_SECONDARY"
                    ]
                  },
                  "title": {
                    "type": "string",
                    "minLength": 1
                  },
                  "subtitle": {
                    "type": "string"
                  },
                  "ctaLabel": {
                    "type": "string"
                  },
                  "ctaLink": {
                    "type": "string",
                    "format": "uri"
                  },
                  "serviceId": {
                    "type": "string"
                  },
                  "mediaId": {
                    "type": "string"
                  },
                  "order": {
                    "type": "integer"
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "placement",
                  "title"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "placement": {
                          "type": "string",
                          "enum": [
                            "HOME_HERO",
                            "CTA_PRIMARY",
                            "CTA_SECONDARY"
                          ]
                        },
                        "title": {
                          "type": "string"
                        },
                        "subtitle": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "ctaLabel": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "ctaLink": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "serviceId": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "order": {
                          "type": "number"
                        },
                        "isActive": {
                          "type": "boolean"
                        },
                        "createdAt": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "placement",
                        "title",
                        "subtitle",
                        "ctaLabel",
                        "ctaLink",
                        "serviceId",
                        "order",
                        "isActive",
                        "createdAt",
                        "updatedAt"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/banners/{id}": {
      "get": {
        "summary": "Get /api/banners/{id}",
        "responses": {
          "200": {
            "description": "Record",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "placement": {
                          "type": "string",
                          "enum": [
                            "HOME_HERO",
                            "CTA_PRIMARY",
                            "CTA_SECONDARY"
                          ]
                        },
                        "title": {
                          "type": "string"
                        },
                        "subtitle": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "ctaLabel": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "ctaLink": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "serviceId": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "order": {
                          "type": "number"
                        },
                        "isActive": {
                          "type": "boolean"
                        },
                        "createdAt": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "placement",
                        "title",
                        "subtitle",
                        "ctaLabel",
                        "ctaLink",
                        "serviceId",
                        "order",
                        "isActive",
                        "createdAt",
                        "updatedAt"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "summary": "Update /api/banners/{id}",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "placement": {
                    "type": "string",
                    "enum": [
                      "HOME_HERO",
                      "CTA_PRIMARY",
                      "CTA_SECONDARY"
                    ]
                  },
                  "title": {
                    "type": "string",
                    "minLength": 1
                  },
                  "subtitle": {
                    "type": "string"
                  },
                  "ctaLabel": {
                    "type": "string"
                  },
                  "ctaLink": {
                    "type": "string",
                    "format": "uri"
                  },
                  "serviceId": {
                    "type": "string"
                  },
                  "mediaId": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "order": {
                    "type": "integer"
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "placement": {
                          "type": "string",
                          "enum": [
                            "HOME_HERO",
                            "CTA_PRIMARY",
                            "CTA_SECONDARY"
                          ]
                        },
                        "title": {
                          "type": "string"
                        },
                        "subtitle": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "ctaLabel": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "ctaLink": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "serviceId": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "order": {
                          "type": "number"
                        },
                        "isActive": {
                          "type": "boolean"
                        },
                        "createdAt": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "placement",
                        "title",
                        "subtitle",
                        "ctaLabel",
                        "ctaLink",
                        "serviceId",
                        "order",
                        "isActive",
                        "createdAt",
                        "updatedAt"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "delete": {
        "summary": "Delete /api/banners/{id}",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/api/settings/{key}": {
      "get": {
        "summary": "Get setting",
        "responses": {
          "200": {
            "description": "Setting",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "key": {
                          "type": "string"
                        },
                        "value": {},
                        "updatedAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "key",
                        "updatedAt"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "summary": "Upsert setting",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "value": {}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Setting updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "key": {
                          "type": "string"
                        },
                        "value": {},
                        "updatedAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "key",
                        "updatedAt"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete setting",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          }
        }
      }
    }
  },
...
}